Update repository patterns
